x86/resctrl: Make rdt_enable_key the arch's decision to switch
ANBZ: #1598

cherry-picked from https://git.kernel.org/pub/scm/linux/kernel/git/morse/linux.git

rdt_enable_key is switched when resctrl is mounted. It was also previously
used to prevent a second mount of the filesystem.

Any other architecture that wants to support resctrl has to provide
identical static keys.

Now that we have helpers for the alloc/mon keys, resctrl doesn't need
to switch this extra key, it can be done by the arch code. To allow
the static key increment/decrement functions to be used, ensure the
calls to the helpers are balanced.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Xin Hao <xhao@linux.alibaba.com>
Tested-by: Shawn Wang <shawnwang@linux.alibaba.com>
Reviewed-by: Baolin Wang <baolin.wang@linux.alibaba.com>