x86/resctrl: Access per-rmid structures by index
ANBZ: #1598

cherry-picked from https://git.kernel.org/pub/scm/linux/kernel/git/morse/linux.git

Because of the differences between RDT and MPAM's monitors, rmid
values on arm64 are not unique unless the closid is also included.
Bitmaps like rmid_busy_llc need to be sized by the number of
unique entries for this resource.

Add helpers to encode/decode the closid+rmid to an index. The
domain's busy_rmid_llc and the rmid_ptrs[] array are then sized by
index. On x86, this is always just the rmid. This gives resctrl a unique
value it can use to store monitor values, and allows MPAM to decode
the closid when reading the hardware counters.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Xin Hao <xhao@linux.alibaba.com>
Tested-by: Shawn Wang <shawnwang@linux.alibaba.com>
Reviewed-by: Baolin Wang <baolin.wang@linux.alibaba.com>