KVM: SVM: Prevent debugging under SEV-ES
ANBZ: #620

commit 8d4846b9b15045598d760470789716fb08b9b317 upstream.

Since the guest register state of an SEV-ES guest is encrypted, debugging
is not supported. Update the code to prevent guest debugging when the
guest has protected state.

Additionally, an SEV-ES guest must only and always intercept DR7 reads and
writes. Update set_dr_intercepts() and clr_dr_intercepts() to account for
this.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>
Message-Id: <8db966fa2f9803d6454ce773863025d0e2e7f3cc.1607620209.git.thomas.lendacky@amd.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: yangge <yangge@hygon.cn>
Reviewed-by: Artie Ding <artie.ding@linux.alibaba.com>