KVM: SVM: Add support for SEV-ES GHCB MSR protocol function 0x100
ANBZ: #620

commit e1d71116b64a54c3948d2692d4338e643408c411 upstream.

The GHCB specification defines a GHCB MSR protocol using the lower
12-bits of the GHCB MSR (in the hypervisor this corresponds to the
GHCB GPA field in the VMCB).

Function 0x100 is a request for termination of the guest. The guest has
encountered some situation for which it has requested to be terminated.
The GHCB MSR value contains the reason for the request.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>
Message-Id: <f3a1f7850c75b6ea4101e15bbb4a3af1a203f1dc.1607620209.git.thomas.lendacky@amd.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: yangge <yangge@hygon.cn>
Reviewed-by: Artie Ding <artie.ding@linux.alibaba.com>