ck: tcp: Add sysctl for TCP_INIT_CWND
to #32811250

The default tcp initial cwnd is 10, that is too small in some
scenarios. But the TCP_INIT_CWND cannot be adjusted for it's
a macro. This patch will add a sysctl for TCP_INIT_CWND and
set it to 10 as the default/min value.

Considering that the TCP_INIT_CWND used to be a rvalue, I
convert the sysctl to rvalue in a tricky way.

However the patch may also have some side effects, which are
not serious but should be mentioned:

1) Some logic in TCP assumes the TCP_INIT_CWND never changes,
   and may compare current snd_cwnd with it. This patch may let
   the logic have wrong results or enter wrong states. But most
   of them should be robust enough to recover from it with new
   ACKs arrived.
2) There're some hard codes in kernel which assume TCP_INIT_CWND
   is always 10. For example, the initial_quantum of sch_fq is 10.
   So some packets may be sent with unexpected sending rate. It's
   ok most of the time, but if we predict the real sending time of
   packets like what have be done in newer kernel, we have to fix
   the hard codes too.

Generally these side effects shouldn't bother you too much, because
the TCP_INIT_CWND shouldn't change frequently. It should be set
on system booting unless you HAVE TO change it after booting.

Signed-off-by: Cambda Zhu <cambda@linux.alibaba.com>
Reviewed-by: Tony Lu <tonylu@linux.alibaba.com>
Acked-by: Dust Li <dust.li@linux.alibaba.com>
Signed-off-by: Qiao Ma <mqaio@linux.alibaba.com>