anolis: virtio-net: bugfix for xsk free the xsk.hdr with race
OpenAnolis Bug Tracker: 0000439

When xsk disable, xsk.hdr will be released, and if the last xsk packet
is recycling, both sides will try to release the xsk.hdr at the same
time. This constitutes a race condition. Therefore, it is necessary to
add a lock in virtnet_xsk_pool_disable for protection. Because of
virt_xsk_complete is under the protection of __netif_tx_lock, so
choose to use __netif_tx_lock in virtnet_xsk_pool_disable to protect.

Signed-off-by: Xuan Zhuo <xuanzhuo@linux.alibaba.com>
Reviewed-by: Dust Li <dust.li@linux.alibaba.com>
Reviewed-by: Shile Zhang <shile.zhang@linux.alibaba.com>