Intel: x86/fpu/xstate: Define the scope of the initial xstate data
to #35358497

commit: https://lore.kernel.org/patchwork/patch/1382797/

init_fpstate is used to record the initial xstate value and covers all the
states. But it is wasteful to cover large states all with trivial initial
data.

Limit init_fpstate by clarifying its size and coverage, which are all but
dynamic user states. The dynamic states are assumed to be large but having
initial data with zeros.

Expand copy_xregs_to_kernel_booting() to receive a mask argument of which
states to save.

[
 Guanjun: fix conflicts in
  arch/x86/include/asm/fpu/internal.h,
  arch/x86/kernel/fpu/xstate.c
]

Signed-off-by: Chang S. Bae <chang.seok.bae@intel.com>
Reviewed-by: Len Brown <len.brown@intel.com>
Cc: x86@kernel.org
Cc: linux-kernel@vger.kernel.org
Signed-off-by: Lin Wang <lin.x.wang@intel.com>
Signed-off-by: Guanjun <guanjun@linux.alibaba.com>
Reviewed-by: Zelin Deng <zelin.deng@linux.alibaba.com>
Acked-by: Artie Ding <artie.ding@linux.alibaba.com>