vfio/mdev: Use struct mdev_type in struct mdev_device
ANBZ: #1087

commit 417fd5bf242d7691c15fe0bd705ab76c69276572 upstream.

The kobj pointer in mdev_device is actually pointing at a struct
mdev_type. Use the proper type so things are understandable.

There are a number of places that are confused and passing both the mdev
and the mtype as function arguments, fix these to derive the mtype
directly from the mdev to remove the redundancy.

Intel-SIG: commit 417fd5bf242d vfio/mdev: Use struct mdev_type in struct mdev_device.
SIOV VDCM feature backporting.

Reviewed-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Kevin Tian <kevin.tian@intel.com>
Reviewed-by: Cornelia Huck <cohuck@redhat.com>
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>
Message-Id: <5-v2-d36939638fc6+d54-vfio2_jgg@nvidia.com>
Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
[ Fengqian Gao: amend commit log ]
Signed-off-by: Fengqian Gao <fengqian.gao@intel.com>
Reviewed-by: Artie Ding <artie.ding@linux.alibaba.com>