vfio/mdev: Reorganize mdev_device_create()
ANBZ: #1087

commit fbd0e2b0c3d0b2eeaef471c9fe19ae5a7b2ee970 upstream.

Once the memory for the struct mdev_device is allocated it should
immediately be device_initialize()'d and filled in so that put_device()
can always be used to undo the allocation.

Place the mdev_get/put_parent() so that they are clearly protecting the
mdev->parent pointer. Move the final put to the release function so that
the lifetime rules are trivial to understand. Update the goto labels to
follow the normal convention.

Remove mdev_device_free() as the release function via device_put() is now
usable in all cases.

Intel-SIG: commit fbd0e2b0c3d0 vfio/mdev: Reorganize mdev_device_create().
SIOV VDCM feature backporting.

Reviewed-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Kevin Tian <kevin.tian@intel.com>
Reviewed-by: Cornelia Huck <cohuck@redhat.com>
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>
Message-Id: <8-v2-d36939638fc6+d54-vfio2_jgg@nvidia.com>
Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
[ Fengqian Gao: amend commit log ]
Signed-off-by: Fengqian Gao <fengqian.gao@intel.com>
Reviewed-by: Artie Ding <artie.ding@linux.alibaba.com>