vfio/mdev: Remove duplicate storage of parent in mdev_device
ANBZ: #1087

commit fbea43239074e16c91048f5ce70378664efbdb99 upstream.

mdev_device->type->parent is the same thing.

The struct mdev_device was relying on the kref on the mdev_parent to also
indirectly hold a kref on the mdev_type pointer. Now that the type holds a
kref on the parent we can directly kref the mdev_type and remove this
implicit relationship.

Intel-SIG: commit fbea43239074 vfio/mdev: Remove duplicate storage of parent in mdev_device.
SIOV VDCM feature backporting.

Reviewed-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Cornelia Huck <cohuck@redhat.com>
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>
Message-Id: <10-v2-d36939638fc6+d54-vfio2_jgg@nvidia.com>
Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
[ Fengqian Gao: amend commit log ]
Signed-off-by: Fengqian Gao <fengqian.gao@intel.com>
Reviewed-by: Artie Ding <artie.ding@linux.alibaba.com>