RISC-V: Add kexec support
commit 6a3a220ab79e6b490a9eb9cf996d65eb617dac2f upstream

This patch adds support for kexec on RISC-V. On SMP systems it depends
on HOTPLUG_CPU in order to be able to bring up all harts after kexec.
It also needs a recent OpenSBI version that supports the HSM extension.
I tested it on riscv64 QEMU on both an smp and a non-smp system.

Signed-off-by: Nick Kossifidis <mick@ics.forth.gr>
Signed-off-by: Palmer Dabbelt <palmerdabbelt@google.com>
Signed-off-by: Baolin Wang <baolin.wang@linux.alibaba.com>