x86/fpu: Move inlines where they belong
ANBZ: #173

commit e68524456c855e500f0a636adb1aa977e1e0b4d8 upstream.

They are only used in fpstate_init() and there is no point to have them in
a header just to make reading the code harder.

Intel-SIG: commit e68524456c85 x86/fpu: Move inlines where they belong.
Intel Advanced Matrix Extensions(AMX) is one of the key features on Sapphire
Rapids(SPR) and it depends on xsave, this patch is part of Thomas Gleixner's
x86 FPU code cleanup that upstream AMX implementation is building on, and is
also key to all code for being developed in upstream that depends on xsave.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Borislav Petkov <bp@suse.de>
Reviewed-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/20210623121452.023118522@linutronix.de
[ Lin Wang: amend commit log ]
Signed-off-by: Lin Wang <lin.x.wang@intel.com>
Reviewed-by: Artie Ding <artie.ding@linux.alibaba.com>