x86/cpu: Sanitize X86_FEATURE_OSPKE
ANBZ: #173

commit 8a1dc55a3f3ef0a723c3c117a567e7b5dd2c1793 upstream.

X86_FEATURE_OSPKE is enabled first on the boot CPU and the feature flag is
set. Secondary CPUs have to enable CR4.PKE as well and set their per CPU
feature flag. That's ineffective because all call sites have checks for
boot_cpu_data.

Make it smarter and force the feature flag when PKU is enabled on the boot
cpu which allows then to use cpu_feature_enabled(X86_FEATURE_OSPKE) all
over the place. That either compiles the code out when PKEY support is
disabled in Kconfig or uses a static_cpu_has() for the feature check which
makes a significant difference in hotpaths, e.g. context switch.

Intel-SIG: commit 8a1dc55a3f3e x86/cpu: Sanitize X86_FEATURE_OSPKE.
Intel Advanced Matrix Extensions(AMX) is one of the key features on Sapphire
Rapids(SPR) and it depends on xsave, this patch is part of Thomas Gleixner's
x86 FPU code cleanup that upstream AMX implementation is building on, and is
also key to all code for being developed in upstream that depends on xsave.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Borislav Petkov <bp@suse.de>
Reviewed-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/20210623121455.305113644@linutronix.de
[ Lin Wang: amend commit log ]
Signed-off-by: Lin Wang <lin.x.wang@intel.com>
Reviewed-by: Artie Ding <artie.ding@linux.alibaba.com>