ck: sched: make ID_LOOSE_EXPEL defined with CONFIG_GROUP_IDENTITY on
fix #34359599
fix #34361563

In commit b7fecff78665 ("ck: sched: fix the performence regression
caused by update_rq_on_expel()") , ID_LOOSE_EXPEL is defined without
#ifdef CONFIG_GROUP_IDENTITY, but actually, it is only effective while
CONFIG_GROUP_IDENTITY is on, and this patch is to fix it.

By the way, the unit of sysctl_sched_expel_update_interval should be
ms, this patch also fix the description and the logic.

Signed-off-by: Cruz Zhao <CruzZhao@linux.alibaba.com>
Acked-by: Michael Wang <yun.wang@linux.alibaba.com>