x86/fpu/signal: Use fpstate for size and features
ANBZ: #173

commit 5509cc78080d29b23706dbf076d51691b69f3c79 upstream.

For dynamically enabled features it's required to get the features which
are enabled for that context when restoring from sigframe.

The same applies for all signal frame size calculations.

Intel-SIG: commit 5509cc78080d x86/fpu/signal: Use fpstate for size and
features.
Move register state into a container struct which carries information about
the fpstate buffer, this is the second part of a 4 part submission which
work towards AMX and dynamic feature enabling.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/87ilxz5iew.ffs@tglx
[ Lin Wang: amend commit log ]
Signed-off-by: Lin Wang <lin.x.wang@intel.com>
Reviewed-by: Artie Ding <artie.ding@linux.alibaba.com>