x86/fpu: Restrict xsaves()/xrstors() to independent states
ANBZ: #173

commit f5daf836f292f795f9cf8f36e036bf47adcbc3a3 upstream.

These interfaces are really only valid for features which are independently
managed and not part of the task context state for various reasons.

Tighten the checks and adjust the misleading comments.

Intel-SIG: commit f5daf836f292 x86/fpu: Restrict xsaves()/xrstors()
to independent states.
Preparatory cleanups for AMX support, this is the first part of a 4 part
submission which work towards AMX and dynamic feature enabling.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/20211015011538.608492174@linutronix.de
[ Lin Wang: amend commit log ]
Signed-off-by: Lin Wang <lin.x.wang@intel.com>
Reviewed-by: Artie Ding <artie.ding@linux.alibaba.com>