x86/fpu: Prepare fpu_clone() for dynamically enabled features
ANBZ: #173

commit 9e798e9aa14c45fb94e47b30bf6347b369ce9df7 upstream.

The default portion of the parent's FPU state is saved in a child task.
With dynamic features enabled, the non-default portion is not saved in a
child's fpstate because these register states are defined to be
caller-saved. The new task's fpstate is therefore the default buffer.

Fork inherits the permission of the parent.

Also, do not use memcpy() when TIF_NEED_FPU_LOAD is set because it is
invalid when the parent has dynamic features.

Intel-SIG: commit 9e798e9aa14c x86/fpu: Prepare fpu_clone() for dynamically
enabled features.
Support Intel Advanced Matrix Extensions(AMX). This is the last part of a
4 part submission which work towards AMX and dynamic feature enabling.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Chang S. Bae <chang.seok.bae@intel.com>
Signed-off-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/20211021225527.10184-11-chang.seok.bae@intel.com
[ Lin Wang: amend commit log ]
Signed-off-by: Lin Wang <lin.x.wang@intel.com>
Reviewed-by: Artie Ding <artie.ding@linux.alibaba.com>