x86/kvm: Convert FPU handling to a single swap buffer
ANBZ: #173

commit d69c1382e1b73a0496a70872a035ca2b22d074e5 upstream.

For the upcoming AMX support it's necessary to do a proper integration with
KVM. Currently KVM allocates two FPU structs which are used for saving the user
state of the vCPU thread and restoring the guest state when entering
vcpu_run() and doing the reverse operation before leaving vcpu_run().

With the new fpstate mechanism this can be reduced to one extra buffer by
swapping the fpstate pointer in current::thread::fpu. This makes the
upcoming support for AMX and XFD simpler because then fpstate information
(features, sizes, xfd) are always consistent and it does not require any
nasty workarounds.

Convert the KVM FPU code over to this new scheme.

Intel-SIG: commit d69c1382e1b7 x86/kvm: Convert FPU handling to a single
swap buffer.
Sanitize the FPU guest/user handling in kvm for the upcoming AMX support.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/20211022185313.019454292@linutronix.de
[ Lin Wang: amend commit log ]
Signed-off-by: Lin Wang <lin.x.wang@intel.com>
Reviewed-by: Artie Ding <artie.ding@linux.alibaba.com>