anolis: kfence: split memblock alloc area into 1GB when system booting
ANBZ: #28

We set the size of 1GB as a basic unit of kfence pool area. To fit this
change, the size of memblock_alloc() is splited to several 1GB areas.
For num_objects_pernode > 65535, we align its pool size up to 1GiB to
use the full pud to keep performance of other space and for
convenience of recovering tlb in future.

Signed-off-by: Tianchen Ding <dtcccc@linux.alibaba.com>
Reviewed-by: Xunlei Pang <xlpang@linux.alibaba.com>