ACPI: CPPC: Add CPPC enable register function
ANBZ: #777

commit fb0b00af04d083770d6e2762b2838357519f7d2d upstream.

Add a new function to enable CPPC feature. This function
will write Continuous Performance Control package
EnableRegister field on the processor.

CPPC EnableRegister register described in section 8.4.7.1 of ACPI 6.4:
This element is optional. If supported, contains a resource descriptor
with a single Register() descriptor that describes a register to which
OSPM writes a One to enable CPPC on this processor. Before this register
is set, the processor will be controlled by legacy mechanisms (ACPI
Pstates, firmware, etc.).

This register will be used for AMD processors to enable AMD P-State
function instead of legacy ACPI P-States.

Signed-off-by: Jinzhou Su <Jinzhou.Su@amd.com>
Signed-off-by: Huang Rui <ray.huang@amd.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
[
 jiayu:
    fix conflicts in
    include/acpi/cppc_acpi.h
]
Signed-off-by: Jiayu Ni <jiayu.ni@linux.alibaba.com>
Reviewed-by: Artie Ding <artie.ding@linux.alibaba.com>