cpufreq: amd-pstate: Introduce the support for the processors with shared memory solution
ANBZ: #777

commit e059c184da47e92c6236f26b0fdaf9e92f0d55b5 upstream.

In some of Zen2 and Zen3 based processors, they are using the shared
memory that exposed from ACPI SBIOS. In this kind of the processors,
there is no MSR support, so we add acpi cppc function as the backend for
them.

It is using a module param (shared_mem) to enable related processors
manually. We will enable this by default once we address performance
issue on this solution.

Signed-off-by: Jinzhou Su <Jinzhou.Su@amd.com>
Signed-off-by: Huang Rui <ray.huang@amd.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Jiayu Ni <jiayu.ni@linux.alibaba.com>
Reviewed-by: Artie Ding <artie.ding@linux.alibaba.com>