x86/fpu: Add guest support to xfd_enable_feature()
ANBZ: #475

commit c270ce393dfd700e7510a4579568deeefba954fd upstream.

Guest support for dynamically enabled FPU features requires a few
modifications to the enablement function which is currently invoked from
the #NM handler:

  1) Use guest permissions and sizes for the update

  2) Update fpu_guest state accordingly

  3) Take into account that the enabling can be triggered either from a
     running guest via XSETBV and MSR_IA32_XFD write emulation or from
     a guest restore. In the latter case the guests fpstate is not the
     current tasks active fpstate.

Split the function and implement the guest mechanics throughout the
callchain.

Compared with upstream patches, there is a modification to
fpstate_realloc in arch/x86/kernel/fpu/xstate.c, that is_confidential
(Indicator for KVM confidential mode) member reference is removed due
to the missing of is_confidential member in struct fpstate.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Jing Liu <jing2.liu@intel.com>
Signed-off-by: Yang Zhong <yang.zhong@intel.com>
Message-Id: <20220105123532.12586-7-yang.zhong@intel.com>
[Add 32-bit stub for __xfd_enable_feature. - Paolo]
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Hao Xiang <hao.xiang@linux.alibaba.com>
Reviewed-by: Artie Ding <artie.ding@linux.alibaba.com>