net/smc: Introduce a new conn->lgr validity check helper
ANBZ: #264

commit ea89c6c0983c39702a4a52ccaa4702e0cb71179b upstream.

It is no longer suitable to identify whether a smc connection
is registered in a link group through checking if conn->lgr
is NULL, because conn->lgr won't be reset even the connection
is unregistered from a link group.

So this patch introduces a new helper smc_conn_lgr_valid() and
replaces all the check of conn->lgr in original implementation
with the new helper to judge if conn->lgr is valid to use.

Signed-off-by: Wen Gu <guwen@linux.alibaba.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Acked-by: Tony Lu <tonylu@linux.alibaba.com>