anolis: net/smc: Keep first contact clcsock
ANBZ: #264

This introduces a work around for eRDMA. eRDMA reuse the first TCP
tuple to create QP, and don't want to release it. But SMC will release
this tuple when first contact connection is shutdown.

This patch keeps the first contact connection, and delay the shutdown
work to link (QP) release progress.

Be careful, this patch reverses TCP close process, which means server
side closes clcsock when link clears.

Signed-off-by: Tony Lu <tonylu@linux.alibaba.com>
Reviewed-by: Wen Gu <guwen@linux.alibaba.com>
Acked-by: Tony Lu <tonylu@linux.alibaba.com>