anolis: x86/cpufeatures: fix {REQUIRED,DISABLED}_MASK_BIT_SET definition.
ANBZ: #1146

The preorder commit increases the value of NCAPINTS, but lacks the
necessary definition modification to {REQUIRED,DISABLED}_MASK_BIT_SET,
so modify it to make it work as expected.

This commit also syncs {required,disabled}-features.h to
tools/arch/x86/include/asm/{required,disabled}-features.h, which is
missing in the original upstream patch.

Fixes: 65a54c154644 ("x86/cpufeatures: x86/msr: Add Intel SGX hardware bits")
Signed-off-by: GuoRui.Yu <GuoRui.Yu@linux.alibaba.com>
Signed-off-by: Zeyu Zhang <zhangzeyu.zzy@alibaba-inc.com>
Acked-by: Zelin Deng <zelin.deng@linux.alibaba.com>